Also apply hass stopping to scripts
